+++ ci4/app/Models/WorkspaceModel.php
@@ -52,17 +52,7 @@ class WorkspaceModel extends Model implements ResourceModelInterface {
             $queryParser->getFilter('status')[0]->ignoreAuto = true;
             $statuses = $queryParser->getFilter('status')[0]->value;
             if (is_array($statuses) && count($statuses) > 0 && $statuses[0] !== '') {
-                $this->groupStart();
-                foreach ($statuses as $status) {
-                    $deploymentStatusSubQuery = (new DeploymentModel())
-                        ->select('COUNT(*) as count', true, false)
-                        ->whereRelated(WorkspaceModel::class, 'id', '${parent}.id', false)
-                        ->where('status', $status)
-                        ->having('count >', 0, true, false);
-
-                    $this->orWhereSubQuery($deploymentStatusSubQuery, '', null, false);
-                }
-                $this->groupEnd();
+                $this->whereIn('status', $statuses);
             }
         }
     }