+++ vue/src/components/Modules/Workspaces/List/WorkspaceList.vue
@@ -1,19 +1,16 @@
 <script setup lang="ts">
-import {computed, defineComponent, onMounted, onUnmounted, reactive, ref, watch} from 'vue'
-import {Api} from "@/core/services/Deploy/Api";
+import { computed, defineComponent, onMounted, onUnmounted, reactive, ref, watch } from "vue";
+import { Api } from "@/core/services/Deploy/Api";
 import bus from "@/plugins/bus";
-import {DeploymentPackage, Workspace} from "@/core/services/Deploy/models";
+import { DeploymentPackage, Workspace } from "@/core/services/Deploy/models";
 import WorkspaceEditButton from "@/components/Modules/Workspaces/EditButton/WorkspaceEditButton.vue";
-import WorkspaceDeploymentStatus
-    from "@/components/Modules/Workspaces/WorkspaceDeploymentStatus/WorkspaceDeploymentStatus.vue";
-import {EventEmitter} from "@/helpers/EventEmitter";
-import debounce from 'lodash.debounce'
-import WorkspaceDeploymentDomains
-    from "@/components/Modules/Workspaces/WorkspaceDeploymentDomains/WorkspaceDeploymentDomains.vue";
-import {it} from "vuetify/locale";
+import WorkspaceDeploymentStatus from "@/components/Modules/Workspaces/WorkspaceDeploymentStatus/WorkspaceDeploymentStatus.vue";
+import { EventEmitter } from "@/helpers/EventEmitter";
+import debounce from "lodash.debounce";
+import WorkspaceDeploymentDomains from "@/components/Modules/Workspaces/WorkspaceDeploymentDomains/WorkspaceDeploymentDomains.vue";
 import AuthService from "@/services/AuthService";
-import {DeploymentStatusTypes, RbacPermissions} from "@/constants";
-import {useRouter} from 'vue-router';
+import { WorkspaceStatusTypes, RbacPermissions } from "@/constants";
+import { useRouter } from "vue-router";
 
 interface Row {
     workspace: Workspace;
@@ -25,11 +22,11 @@ const router = useRouter();
 const itemCount = ref(0);
 const rows = ref<Row[]>([]);
 const headers = ref([
-    {title: 'Name', key: 'workspace.name_readable', sortable: false},
-    {title: 'Namespace', key: 'workspace.namespace', sortable: false},
-    {title: 'Status', key: 'status', sortable: false},
-    {title: 'Url', key: 'url', sortable: false},
-    {title: '', key: 'actions', sortable: false},
+    { title: "Name", key: "workspace.name_readable", sortable: false },
+    { title: "Namespace", key: "workspace.namespace", sortable: false },
+    { title: "Status", key: "status", sortable: false },
+    { title: "Url", key: "url", sortable: false },
+    { title: "", key: "actions", sortable: false },
 ]);
 const isLoading = ref(true);
 const options = ref({});
@@ -38,30 +35,30 @@ const showCreateMenu = ref(false);
 const deploymentPackages = ref<DeploymentPackage[]>([]);
 const showDeploymentPackagesWarning = ref(true);
 
-const searchValue = ref('');
+const searchValue = ref("");
 const statusOptions = ref([
     {
-        value: DeploymentStatusTypes.Draft,
-        title: 'Draft',
+        value: WorkspaceStatusTypes.Draft,
+        title: "Draft",
     },
     {
-        value: DeploymentStatusTypes.Deploying,
-        title: 'Deploying',
+        value: WorkspaceStatusTypes.Deploying,
+        title: "Deploying",
     },
     {
-        value: DeploymentStatusTypes.Active,
-        title: 'Active',
+        value: WorkspaceStatusTypes.Active,
+        title: "Active",
     },
     {
-        value: DeploymentStatusTypes.Error,
-        title: 'Error',
+        value: WorkspaceStatusTypes.Inactive,
+        title: "Inactive",
+    },
+    {
+        value: WorkspaceStatusTypes.Error,
+        title: "Error",
     },
 ]);
-const selectedStatus = ref([
-    DeploymentStatusTypes.Deploying,
-    DeploymentStatusTypes.Active,
-    DeploymentStatusTypes.Error,
-])
+const selectedStatus = ref([WorkspaceStatusTypes.Deploying, WorkspaceStatusTypes.Active, WorkspaceStatusTypes.Error]);
 
 const rbacDeveloper = ref(false);
 const rbacWorkspaceCreate = ref(false);
@@ -74,27 +71,34 @@ onMounted(() => {
     rbacWorkspaceUpdate.value = AuthService.currentAuthUser?.hasPermission(RbacPermissions.Workspaces.Update) ?? false;
     rbacWorkspaceDelete.value = AuthService.currentAuthUser?.hasPermission(RbacPermissions.Workspaces.Delete) ?? false;
 
-    bus.on('workspaceSaved', onItemSaved);
+    bus.on("workspaceSaved", onItemSaved);
 
     getItems(false, true);
 
-    Api.deploymentPackages().get()
-        .find(items => {
+    Api.deploymentPackages()
+        .get()
+        .find((items) => {
             deploymentPackages.value = items;
             showDeploymentPackagesWarning.value = deploymentPackages.value.length === 0;
         });
 });
 
 onUnmounted(() => {
-    bus.off('workspaceSaved', onItemSaved);
+    bus.off("workspaceSaved", onItemSaved);
 });
 
-watch(searchValue, debounce(() => {
-    getItems(true, true);
-}, 500));
-watch(selectedStatus, debounce(() => {
-    getItems(true, true);
-}, 500));
+watch(
+    searchValue,
+    debounce(() => {
+        getItems(true, true);
+    }, 500)
+);
+watch(
+    selectedStatus,
+    debounce(() => {
+        getItems(true, true);
+    }, 500)
+);
 
 function onItemSaved() {
     getItems(true, true);
@@ -111,25 +115,23 @@ function getItems(doItems = true, doCount = false) {
     const api = Api.workspaces().get();
 
     if (searchValue.value?.length) {
-        api
-            .search('aliases', searchValue.value)
-            .search('name_readable', searchValue.value)
-            .search('name_system', searchValue.value)
-            .search('namespace', searchValue.value)
-            .search('subdomain', searchValue.value);
+        api.search("aliases", searchValue.value)
+            .search("name_readable", searchValue.value)
+            .search("name_system", searchValue.value)
+            .search("namespace", searchValue.value)
+            .search("subdomain", searchValue.value);
     }
 
-    api.whereIn('status', selectedStatus.value);
+    api.whereIn("status", selectedStatus.value);
 
     if (doItems) {
-        api
-            .include('deployment')
-            .include('domain')
+        api.include("deployment")
+            .include("domain")
             .limit(tableOptions.itemsPerPage)
             .offset(tableOptions.itemsPerPage * (tableOptions.page - 1))
-            .orderAsc('name_readable')
-            .find(items => {
-                rows.value = items.map(item => {
+            .orderAsc("name_readable")
+            .find((items) => {
+                rows.value = items.map((item) => {
                     return {
                         workspace: item,
                         isLoadingRequestSupportLogin: false,
@@ -141,7 +143,7 @@ function getItems(doItems = true, doCount = false) {
 
     // Count total amount of items
     if (doCount) {
-        api.count(count => {
+        api.count((count) => {
             itemCount.value = count;
         });
     }
@@ -150,31 +152,30 @@ function getItems(doItems = true, doCount = false) {
 // <editor-fold desc="View function">
 
 function onCreateItemBtnClicked(type: DeploymentPackage) {
-    bus.emit('workspaceCreate', {
+    bus.emit("workspaceCreate", {
         deploymentPackage: type,
     });
 }
 
 function onDeployItemBtnClicked(item: Workspace) {
-    bus.emit('confirm', {
+    bus.emit("confirm", {
         body: `Do you want to deploy <strong>${item.name}</strong>?`,
-        confirmIcon: 'fa fa-play',
-        confirmColor: 'green',
+        confirmIcon: "fa fa-play",
+        confirmColor: "green",
 
         responseCallback: (confirmed: boolean) => {
             if (confirmed) {
                 const workerProps = {
                     title: `Deploying ${item.name}`,
-                    body: 'This may take a minute',
-                    onFinishBody: 'All done',
-                    onIsWorkingChangeEventEmitter: new EventEmitter<boolean>()
+                    body: "This may take a minute",
+                    onFinishBody: "All done",
+                    onIsWorkingChangeEventEmitter: new EventEmitter<boolean>(),
                 };
-                bus.emit('worker', workerProps);
+                bus.emit("worker", workerProps);
 
                 workerProps.onIsWorkingChangeEventEmitter.emit(true);
-                const api = Api.workspaces()
-                    .deployPutById(item.id!);
-                api.setErrorHandler(response => {
+                const api = Api.workspaces().deployPutById(item.id!);
+                api.setErrorHandler((response) => {
                     if (response.error) {
                         workerProps.onFinishBody = response.error.replaceAll("\n", "<br>");
                     }
@@ -182,33 +183,32 @@ function onDeployItemBtnClicked(item: Workspace) {
                 });
                 api.save(null, () => {
                     workerProps.onIsWorkingChangeEventEmitter.emit(false);
-                    bus.emit('workspaceSaved');
+                    bus.emit("workspaceSaved");
                 });
             }
-        }
+        },
     });
 }
 
 function onTerminateItemBtnClicked(item: Workspace) {
-    bus.emit('confirm', {
+    bus.emit("confirm", {
         body: `Do you want to terminate <strong>${item.name}</strong>?`,
-        confirmIcon: 'fa fa-skull',
-        confirmColor: 'red',
+        confirmIcon: "fa fa-skull",
+        confirmColor: "red",
 
         responseCallback: (confirmed: boolean) => {
             if (confirmed) {
                 const workerProps = {
                     title: `Terminating ${item.name}`,
-                    body: 'This may take a minute',
-                    onFinishBody: 'All done',
-                    onIsWorkingChangeEventEmitter: new EventEmitter<boolean>()
+                    body: "This may take a minute",
+                    onFinishBody: "All done",
+                    onIsWorkingChangeEventEmitter: new EventEmitter<boolean>(),
                 };
-                bus.emit('worker', workerProps);
+                bus.emit("worker", workerProps);
 
                 workerProps.onIsWorkingChangeEventEmitter.emit(true);
-                const api = Api.workspaces()
-                    .terminatePutById(item.id!);
-                api.setErrorHandler(response => {
+                const api = Api.workspaces().terminatePutById(item.id!);
+                api.setErrorHandler((response) => {
                     if (response.error) {
                         workerProps.onFinishBody = response.error.replaceAll("\n", "<br>");
                     }
@@ -216,86 +216,75 @@ function onTerminateItemBtnClicked(item: Workspace) {
                 });
                 api.save(null, () => {
                     workerProps.onIsWorkingChangeEventEmitter.emit(false);
-                    bus.emit('workspaceSaved');
+                    bus.emit("workspaceSaved");
                 });
             }
-        }
+        },
     });
 }
 
 function onDeleteItemBtnClicked(item: Row) {
-    bus.emit('confirm', {
+    bus.emit("confirm", {
         body: `Do you want to delete <strong>${item.workspace.name}</strong>?`,
-        confirmIcon: 'fa fa-trash',
-        confirmColor: 'red',
+        confirmIcon: "fa fa-trash",
+        confirmColor: "red",
 
         responseCallback: (confirmed: boolean) => {
             if (confirmed) {
                 item.isLoadingDeleteBtn = true;
-                Api.workspaces().get()
-                    .where('id', item.workspace.id!)
-                    .include('deployment')
-                    .find(workspaces => {
+                Api.workspaces()
+                    .get()
+                    .where("id", item.workspace.id!)
+                    .include("deployment")
+                    .find((workspaces) => {
                         item.isLoadingDeleteBtn = false;
 
                         if (workspaces[0].deployments?.length) {
-                            bus.emit('toast', {
-                                text: 'All deployments must be terminated and deleted before you can delete the workspace'
+                            bus.emit("toast", {
+                                text: "All deployments must be terminated and deleted before you can delete the workspace",
                             });
                         } else {
-                            Api.workspaces().deleteById(item.workspace.id!).delete(() => bus.emit('workspaceSaved'));
+                            Api.workspaces()
+                                .deleteById(item.workspace.id!)
+                                .delete(() => bus.emit("workspaceSaved"));
                         }
                     });
             }
-        }
+        },
     });
 }
 
 function onShowMigrationJobsBtnClicked(item: Workspace) {
-    bus.emit('migrationJobList', {
+    bus.emit("migrationJobList", {
         workspace: item,
     });
 }
 
 function onShowDeploymentsBtnClicked(item: Workspace) {
-    bus.emit('workspaceDeploymentList', {
+    bus.emit("workspaceDeploymentList", {
         workspace: item,
     });
 }
 
 function onShowLogsBtnClicked(item: Workspace) {
-    bus.emit('workspaceLogs', {
+    bus.emit("workspaceLogs", {
         workspace: item,
     });
 }
 
 function onDeploymentPackagesShortcutClicked() {
-    router.push({name: 'DeploymentPackages'}).catch((e: any) => {
-    });
+    router.push({ name: "DeploymentPackages" }).catch((e: any) => {});
 }
 
 // </editor-fold>
-
 </script>
 
 <template>
     <div class="h-100 content-wrapper">
-
-        <v-toolbar
-            density="compact"
-            flat
-            color="blue-grey lighten-5"
-            dark
-            height="120"
-        >
-            <div
-                class="d-flex flex-column w-100 py-2 px-4 gap-1"
-            >
+        <v-toolbar density="compact" flat color="blue-grey lighten-5" dark height="120">
+            <div class="d-flex flex-column w-100 py-2 px-4 gap-1">
                 <div class="d-flex">
-
-                    <v-toolbar-title
-                        class="my-auto"
-                    >Workspaces</v-toolbar-title>
+                    <v-toolbar-title class="my-auto">Workspaces</v-toolbar-title>
 
                     <v-spacer></v-spacer>
 
@@ -305,29 +294,19 @@ function onDeploymentPackagesShortcutClicked() {
                         :close-on-content-click="false"
                         left
                         min-width="250"
-                        offset-y>
+                        offset-y
+                    >
                         <template v-slot:activator="{ props }">
-                            <v-btn
-                                v-bind="props"
-                                small
-                                prepend-icon="fa fa-plus"
-                                style="margin-right: -16px"
-                            >
-                                Create
-                            </v-btn>
+                            <v-btn v-bind="props" small prepend-icon="fa fa-plus" style="margin-right: -16px"> Create </v-btn>
                         </template>
 
-                        <v-list
-                            v-if="showDeploymentPackagesWarning"
-                            class="list-items">
-                            <v-list-item
-                                dense>
+                        <v-list v-if="showDeploymentPackagesWarning" class="list-items">
+                            <v-list-item dense>
                                 <v-list-item-title>
                                     <span class="font-italic">No Workspace Template found.</span>
                                 </v-list-item-title>
                             </v-list-item>
-                            <v-list-item
-                                @click="onDeploymentPackagesShortcutClicked">
+                            <v-list-item @click="onDeploymentPackagesShortcutClicked">
                                 <v-list-item-title>
                                     <v-icon size="small" class="my-auto">fa fa-circle-right</v-icon>
                                     <span class="ml-2">Go to Workspace Templates</span>
@@ -335,13 +314,8 @@ function onDeploymentPackagesShortcutClicked() {
                             </v-list-item>
                         </v-list>
 
-                        <v-list
-                            v-else
-                            class="list-items">
-                            <v-list-item
-                                v-for="(type, i) in deploymentPackages" :key="i"
-                                dense
-                                @click="onCreateItemBtnClicked(type)">
+                        <v-list v-else class="list-items">
+                            <v-list-item v-for="(type, i) in deploymentPackages" :key="i" dense @click="onCreateItemBtnClicked(type)">
                                 <v-list-item-title>
                                     <v-icon size="small" class="my-auto">fa fa-window-maximize fa</v-icon>
                                     <span class="ml-2">{{ type.name }}</span>
@@ -351,10 +325,7 @@ function onDeploymentPackagesShortcutClicked() {
                     </v-menu>
                 </div>
 
-                <div
-                    class="d-flex gap-1"
-                >
-
+                <div class="d-flex gap-1">
                     <v-text-field
                         v-model="searchValue"
                         variant="outlined"
@@ -378,7 +349,6 @@ function onDeploymentPackagesShortcutClicked() {
                         closable-chips
                         clearable
                     />
-
                 </div>
             </div>
         </v-toolbar>
@@ -391,80 +361,97 @@ function onDeploymentPackagesShortcutClicked() {
             :items-per-page="-1"
             class="table"
             density="compact"
-            @update:options="options = $event; getItems()">
-
+            @update:options="
+                options = $event;
+                getItems();
+            "
+        >
             <template v-slot:item.status="{ item }">
-                <WorkspaceDeploymentStatus
-                    :workspace="item.workspace"/>
+                <WorkspaceDeploymentStatus :workspace="item.workspace" />
             </template>
 
             <template v-slot:item.url="{ item }">
-                <workspace-deployment-domains
-                    :workspace="item.workspace"
-                />
+                <workspace-deployment-domains :workspace="item.workspace" />
             </template>
 
             <template v-slot:item.actions="{ item }">
                 <div class="d-flex justify-end">
-
                     <v-btn
                         v-if="rbacDeveloper"
-                        variant="plain" color="primary" size="small" icon
-                        @click="onShowDeploymentsBtnClicked(item.workspace)">
+                        variant="plain"
+                        color="primary"
+                        size="small"
+                        icon
+                        @click="onShowDeploymentsBtnClicked(item.workspace)"
+                    >
                         <v-icon>fa fa-box</v-icon>
                         <v-tooltip activator="parent" location="bottom">Deployments</v-tooltip>
                     </v-btn>
 
                     <v-btn
                         v-if="rbacDeveloper"
-                        variant="plain" color="primary" size="small" icon
-                        @click="onShowLogsBtnClicked(item.workspace)">
+                        variant="plain"
+                        color="primary"
+                        size="small"
+                        icon
+                        @click="onShowLogsBtnClicked(item.workspace)"
+                    >
                         <v-icon>fa fa-rectangle-list</v-icon>
                         <v-tooltip activator="parent" location="bottom">Kubernetes Logs</v-tooltip>
                     </v-btn>
 
                     <v-btn
                         v-if="rbacDeveloper"
-                        variant="plain" color="primary" size="small" icon
-                        @click="onShowMigrationJobsBtnClicked(item.workspace)">
+                        variant="plain"
+                        color="primary"
+                        size="small"
+                        icon
+                        @click="onShowMigrationJobsBtnClicked(item.workspace)"
+                    >
                         <v-icon>fa fa-truck-arrow-right</v-icon>
                         <v-tooltip activator="parent" location="bottom">Migration Jobs</v-tooltip>
                     </v-btn>
 
-                    <v-menu
-                        v-if="rbacWorkspaceUpdate"
-                        min-width="250">
+                    <v-menu v-if="rbacWorkspaceUpdate" min-width="250">
                         <template v-slot:activator="{ props }">
-                            <v-btn
-                                v-bind="props"
-                                variant="plain" color="primary" size="small" icon>
+                            <v-btn v-bind="props" variant="plain" color="primary" size="small" icon>
                                 <v-icon>fa fa-cog</v-icon>
                                 <v-tooltip activator="parent" location="bottom">Settings</v-tooltip>
                             </v-btn>
                         </template>
-                        <WorkspaceEditButton
-                            :workspace="item.workspace"/>
+                        <WorkspaceEditButton :workspace="item.workspace" />
                     </v-menu>
 
                     <v-btn
                         v-if="rbacWorkspaceCreate"
-                        variant="plain" color="warning" size="small" icon
-                        @click="onDeployItemBtnClicked(item.workspace)">
+                        variant="plain"
+                        color="warning"
+                        size="small"
+                        icon
+                        @click="onDeployItemBtnClicked(item.workspace)"
+                    >
                         <v-icon>fa fa-play</v-icon>
                         <v-tooltip activator="parent" location="bottom">Deploy</v-tooltip>
                     </v-btn>
 
                     <v-btn
                         v-if="rbacDeveloper"
-                        variant="plain" color="red" size="small" icon
-                        @click="onTerminateItemBtnClicked(item.workspace)">
+                        variant="plain"
+                        color="red"
+                        size="small"
+                        icon
+                        @click="onTerminateItemBtnClicked(item.workspace)"
+                    >
                         <v-icon>fa fa-skull</v-icon>
                         <v-tooltip activator="parent" location="bottom">Terminate</v-tooltip>
                     </v-btn>
 
                     <v-btn
                         v-if="rbacWorkspaceUpdate"
-                        variant="plain" color="red" size="small" icon
+                        variant="plain"
+                        color="red"
+                        size="small"
+                        icon
                         @click="onDeleteItemBtnClicked(item)"
                         :loading="item.isLoadingDeleteBtn"
                     >